src/azstorage-upload/requirements.txt
-i https://pypi.org/simple
azure-core==1.5.0
azure-identity==1.3.1
azure-storage-blob==12.3.1
certifi==2020.4.5.1
cffi==1.14.0
chardet==3.0.4
click==7.1.2
cryptography==2.9.2
flask==1.1.2
git+https://github.com/rayluo/flask-session.git@ea8351438e4c79bad59429a43dd662c167c32a32#egg=flask-session
gunicorn==20.0.4
idna==2.9
isodate==0.6.0
itsdangerous==1.1.0
jinja2==2.11.2
markupsafe==1.1.1
msal-extensions==0.1.3
msal==1.2.0
msrest==0.6.13
oauthlib==3.1.0
portalocker==1.7.0
pycparser==2.20
pyjwt[crypto]==1.7.1
requests-oauthlib==1.3.0
requests==2.23.0
six==1.14.0
urllib3==1.25.9
werkzeug==1.0.1
